fix slight spacing inconsistency
Adding margin to the top-content div in phone size exposed more of the
banner image instead of showing padding.

Placing a background-color on .container makes it so adding margin to
the top nav at the phone breakpoint pushes the image down while exposing
more of .container's background (which is now opaque instead of
transparent).

This has the desired result.